# services/auth/openapi.yaml
openapi: 3.0.3
info:
  title: PeoplePay Authentication Service
  description: Authentication service with OAuth2, biometric login, and 2FA
  version: 1.0.0
  contact:
    name: PeoplePay Team
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://auth.peoplepay.com
    description: Production server

paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  service:
                    type: string
                    example: auth
                  timestamp:
                    type: string
                    format: date-time

  /auth/signup:
    post:
      summary: User registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                phoneNumber:
                  type: string
      responses:
        '200':
          description: Registration successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  tokens:
                    $ref: '#/components/schemas/AuthTokens'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/login:
    post:
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                userAgent:
                  type: string
                ip:
                  type: string
                fingerprint:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  tokens:
                    $ref: '#/components/schemas/AuthTokens'
                  requiresMFA:
                    type: boolean
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/refresh:
    post:
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '401':
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/2fa/setup:
    post:
      summary: Setup two-factor authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
              properties:
                userId:
                  type: string
      responses:
        '200':
          description: 2FA setup initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  secret:
                    type: string
                  qrCode:
                    type: string
        '400':
          description: Setup failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/2fa/verify:
    post:
      summary: Verify two-factor authentication code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
                - code
              properties:
                userId:
                  type: string
                code:
                  type: string
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema:
                type: object
                properties:
                  verified:
                    type: boolean
        '400':
          description: Verification failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/biometric/setup:
    post:
      summary: Setup biometric authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
              properties:
                userId:
                  type: string
      responses:
        '200':
          description: Biometric setup options
          content:
            application/json:
              schema:
                type: object
                description: WebAuthn registration options
        '400':
          description: Setup failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/biometric/verify-setup:
    post:
      summary: Verify biometric setup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
                - response
              properties:
                userId:
                  type: string
                response:
                  type: object
                  description: WebAuthn registration response
      responses:
        '200':
          description: Setup verification result
          content:
            application/json:
              schema:
                type: object
                properties:
                  verified:
                    type: boolean
        '400':
          description: Verification failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/biometric/authenticate:
    post:
      summary: Initiate biometric authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
              properties:
                userId:
                  type: string
      responses:
        '200':
          description: Biometric authentication options
          content:
            application/json:
              schema:
                type: object
                description: WebAuthn authentication options
        '400':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/biometric/verify-auth:
    post:
      summary: Verify biometric authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
                - response
              properties:
                userId:
                  type: string
                response:
                  type: object
                  description: WebAuthn authentication response
      responses:
        '200':
          description: Authentication verification result
          content:
            application/json:
              schema:
                type: object
                properties:
                  verified:
                    type: boolean
        '400':
          description: Verification failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /oauth/authorize:
    get:
      summary: OAuth2 authorization endpoint
      parameters:
        - name: client_id
          in: query
          required: true
          schema:
            type: string
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
        - name: code_challenge
          in: query
          required: true
          schema:
            type: string
        - name: code_challenge_method
          in: query
          schema:
            type: string
            enum: [S256, plain]
            default: S256
        - name: state
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Authorization initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  authorizationUrl:
                    type: string
                  state:
                    type: string
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /oauth/token:
    post:
      summary: OAuth2 token endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - grant_type
                - code
                - code_verifier
                - client_id
                - redirect_uri
              properties:
                grant_type:
                  type: string
                  enum: [authorization_code]
                code:
                  type: string
                code_verifier:
                  type: string
                client_id:
                  type: string
                redirect_uri:
                  type: string
      responses:
        '200':
          description: Token issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
                    example: Bearer
                  expires_in:
                    type: integer
                  refresh_token:
                    type: string
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/sessions:
    get:
      summary: Get active user sessions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Active sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Session'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/sessions/{sessionId}:
    delete:
      summary: Revoke a user session
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        passwordHash:
          type: string
        phoneNumber:
          type: string
        kycLevel:
          type: string
          enum: [basic, verified, premium]
        twoFactorEnabled:
          type: boolean
        biometricEnabled:
          type: boolean
        twoFactorSecret:
          type: string
        biometricCredentialId:
          type: string
        biometricPublicKey:
          type: string
        sessionId:
          type: string
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time

    AuthTokens:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer

    Session:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        deviceInfo:
          type: object
          properties:
            userAgent:
              type: string
            ip:
              type: string
            fingerprint:
              type: string
        createdAt:
          type: string
          format: date-time
        lastActivityAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        isActive:
          type: boolean

    Error:
      type: object
      properties:
        error:
          type: string

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT